requirements.txt
python-dotenv~=1.0.1
fastapi~=0.115.4
uvicorn~=0.32.0
aioredis~=2.0.1
requests~=2.32.3
pydantic~=2.9.2
Booktype
websocket-client~=1.8.0
httpx~=0.27.2
websocket~=0.2.1
websockets~=14.0
web3~=7.5.0
sqlparse~=0.5.1
pytest~=8.3.3